Change cache size data type from uint32 -> unit64

+++ internal/cache/cache.go
@@ -54,7 +54,7 @@ func (c *RistrettoCache) Get(key string) (val CacheEntry, hit bool) {
 	return val, hit;
 }
 
-func NewRistrettoCache(cacheSize uint32) *RistrettoCache {
+func NewRistrettoCache(cacheSize uint64) *RistrettoCache {
 	/**  Maxcost and NumCounters
 	 *
 	 * This Ristretto cache is configured with a max size in bytes (the
@@ -109,7 +109,7 @@ func NewNoCache() *NoCache {
  *  like a cache, but does not cache anything. I.e. it will always be empty and
  *  will have a 100% cache misses.
  */
-func NewCache(cachesize uint32) Cache {
+func NewCache(cachesize uint64) Cache {
 	if cachesize == 0 {
 		return NewNoCache()
 	}
